+++ include/oif/globals.h
@@ -1,6 +1,2 @@
 #pragma once
 
-enum {
-    OIF_ERROR = 101,
-    OIF_BACKEND_INIT_ERROR = 102,
-};
