+++ src/oif/backend_c/backend_c.c
@@ -5,7 +5,7 @@
 #include <string.h>
 
 #include <oif/api.h>
-#include <oif/backend_api.h>
+#include <oif/dispatch_api.h>
 #include <oif/dispatch.h>
